fix speedlimit for iplc again

+++ service/controller/config.go
@@ -8,7 +8,7 @@ type Config struct {
 	EnableDNS            bool              `mapstructure:"EnableDNS"`
 	DNSType              string            `mapstructure:"DNSType"`
 	IPLCCheckDuration    int               `mapstructure:"IPLCCheckDuration"`    // minute
-	IPLCSpeedLimit       int64             `mapstructure:"IPLCSpeedLimit"`       // mbps
+	IPLCSpeedLimit       int               `mapstructure:"IPLCSpeedLimit"`       // mbps
 	IPLCSilentSpeedLimit uint64            `mapstructure:"IPLCSilentSpeedLimit"` // mbps
 	IPLCSilentDuration   int               `mapstructure:"IPLCSilentDuration"`   // minute
 	DisableUploadTraffic bool              `mapstructure:"DisableUploadTraffic"`
